pyproject.toml
[project]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Ray da Silva Basilio",email = "58826286+RayBasilio123@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-google-genai (>=2.1.2,<3.0.0)",
    "langgraph (>=0.3.25,<0.4.0)",
    "langchain-anthropic (>=0.1.0,<1.0.0)",
    "langchain-openai (>=0.3.12,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
